+++ glparchis/ui/frmInitGame.py
@@ -1,7 +1,7 @@
 from random import seed
 from datetime import datetime
-from PyQt5.QtCore import Qt
-from PyQt5.QtWidgets import QWizard
+from PyQt6.QtCore import Qt
+from PyQt6.QtWidgets import QWizard
 from glparchis.ui.wdgPlayer import wdgPlayer, SetWdgPlayers
 from glparchis.ui.Ui_frmInitGame import Ui_frmInitGame
 
@@ -18,7 +18,7 @@ class frmInitGame(QWizard, Ui_frmInitGame):
         self.wizardPage2.setTitle(self.tr("Elegir el jugador que empieza la partida"))
         self.wizardPage2.setSubTitle(self.tr("El jugador que saque la puntuacion mas alta, empieza la partida"))        
         
-        self.setButtonText(QWizard.FinishButton, self.tr("Los jugadores lanzan sus dados"))
+        self.setButtonText(QWizard.WizardButton.FinishButton, self.tr("Los jugadores lanzan sus dados"))
         seed(datetime.now().microsecond)
         self.players=[]#players primera del wizard
         self.set=None #SetWdgPlayers()#Set para tirar dados un turno
@@ -32,7 +32,7 @@ class frmInitGame(QWizard, Ui_frmInitGame):
                 
         #Pone juega ordenador en todos menos el primero
         for i in range (1, self.mem.maxplayers):
-            self.players[i].chkIA.setCheckState(Qt.Checked)
+            self.players[i].chkIA.setCheckState(Qt.CheckState.Checked)
         
 
     def validateCurrentPage(self):
@@ -43,7 +43,7 @@ class frmInitGame(QWizard, Ui_frmInitGame):
             for i,  w in enumerate(self.players):
                 w.commit()#Rellena objeto jugador
                 self.mem.settings.setValue("Players/{}".format(w.jugador.color.name), w.jugador.name)
-                if w.chkPlays.checkState()==Qt.Unchecked:#Quita los que no juegan
+                if w.chkPlays.checkState() == Qt.CheckState.Unchecked: #Quita los que no juegan
                     continue
                 self.set.arr.append(w)
             self.scrollPlayerDado.addWidget(self.set.qwidget())
@@ -58,7 +58,7 @@ class frmInitGame(QWizard, Ui_frmInitGame):
                     self.mem.jugadores.actual=maxplayers.arr[0].jugador
                     self.lblPlayerStarts.setStyleSheet('QLabel {font: bold ; color: '+ self.mem.jugadores.actual.color.name+';}')
                     self.lblPlayerStarts.setText(self.tr("El jugador {0} empieza la partida".format(self.mem.jugadores.actual.name)))
-                    self.setButtonText(QWizard.FinishButton, self.tr("Empieza la partida"))          
+                    self.setButtonText(QWizard.WizardButton.FinishButton, self.tr("Empieza la partida"))          
                     return False#Debe haber la validaci´on con playerstarts
                 else:#Mas de dos jugadores
                     self.lblPlayerStarts.setText(self.tr("Los jugadores {} deben tirar hasta que se aclare quien empieza la partida".format(maxplayers.strNames())))